requirements.txt
fastapi
uvicorn
httpx
google-generativeai
pydantic
python-dotenv
pytest
